+++ lib/string_collection_input/string_collection_input_event.dart
@@ -1,0 +1,13 @@
+abstract class StringCollectionInputEvent {}
+
+class ItemAddedEvent extends StringCollectionInputEvent {
+  ItemAddedEvent(this.item);
+
+  final String item;
+}
+
+class ItemRemovedEvent extends StringCollectionInputEvent {
+  ItemRemovedEvent(this.item);
+
+  final String item;
+}