+++ CHANGELOG.md
@@ -12,8 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Firebase account creation and authentication
 - Retrieval of locations, climbs and attempts from firebase
 - Views for locations and climbs
-- Location, climb, attempt, sections and grade set creation
+- Location, climb, attempt, categories and grade set creation
 - Location and climb editing
-- Location, climb and attempt deletion
+- Location, climb, attempt and category deletion
 - Climb filtering
 - App icon